requirements.txt
Flask==1.0.2
ontodev-gizmos
